+++ CHANGES.rst
@@ -17,6 +17,7 @@ Changes
 - add py.typed (don't know why I didn't before...)
 - get rid of :class:`HandlingEndpointMeta` in favor of :class:`.Dispatcher`
 - :func:`event` was moved to :mod:`decorator`
+- :class:`.Dispatcher` was moved to :mod:`handler`
 
 
 Version 0.0.5